update on the add on module
The method peaks (which calculates H infinity norms for S and T given L) was kept.

New function plot_slope was added. It plots the slope of the logarithmic magnitude of a transfer function (like Skogestad 2.10).

A new method called cross_over_freq was made which does the same thing as margin but only returns the first cross over frequency. It was developed by mistake... However, it does throw an interesting exception when I try to close a figure window and this exception is partly why I didn't delete it... Do you know why it happens?