Kalman Controllable and Observable form functions
I implemented Kalman decompositions for the controllable form and the
observable form which can be used sequentially to obtain a minimal
realisation.  This method makes use of the QR decomposition of the
controllability and observability matrices. I noticed that there is
another minimal realisation method that has been implemented, however it
makes use of a more convoluted method.