Reworking of directory structure to package the code as robustcontrol, and initial setup files to do local PIP install (commit 9)
Reworked: Moved the contents of the robustcontrol directory up one level so that the root repository functions as the main directory of the package. Edited the .travis.yml to pip install the package and run test/test_all_notebooks.py. Original commit: Changes were made to add the robustcontrol directory and move the module .py files as suggested on the wiki. The textbook reproductions have also been moved. After doing a PIP install of the package further changes were made to all the reproduction .ipynb files to allow for importing methods from the package. The file "test_all_notebooks.py" and ".testignore" were moved to the robustcontrol/tests directory and run through pytest to verify the changes that were made.